requirements.txt
torch==1.12.1
torchvision==0.13.1
streamlit==1.13.0
Pillow==9.1.0
requests==2.28.1
pandas==1.4.2
matplotlib==3.5.2
seaborn==0.11.2
scikit-learn==1.1.1
